add disabled incomplete support for building libboda.so to makefile.
builds, and resulting lib can be linked with an empty object file to make a working executable. however, to be useful, we'd
minimally need to partition boda into main/everthing else so that the lib could be linked into another program. however,
this was meant just as a sanity check to see if an .so file would be buildable, and shake out some bugs wrt that. in
particular, -fPIC was added to CPPFLAGS in the makefile, and libturbojpeg was added to the (implicit) list of build
dependancy packages. prior to this test, the build was using a static/non-PIC version of libturbojpeg, which was okay for
building an exe, but not for building an .so