cap_app/util: add debug option and printouts for debugging dropped frames.
note that currently everthing seems to be 'as expected', but that
doesn't mean it's good: in the worst case, if proc updates are at the
same rate as the frame updates, long sequences of captured frames can
be dropped/non-displayed, even though all frames are read from the
camera *and* the GUI/display is never stuck. neat! anyway, N-way
buffer sharing should make the curent in_redisplay handling go away so
then it would be moot. otherwise we could probably fix/change things
to give a better result.