V4L2/epoll() bug fixed after upgrading kernel to 3.13, remove polling workaround
since ubuntu 12.04 users can reasonably upgrade to the 14.04 HWE
stack, for now at least we'll remove the polling-based workaround in
cs_disp. that way we can avoid needing to replicate (or factor out and
manage) the workaround globally. note that running 'boda capture' and
checking for any read_but_dropped_frames (should be 0) seems to be an
easy and robust way to check for the issue.