use pos arg support to enhance/fix boda bash completion support.
i think it's now at least not worse than no support. the mode names
are completed as before, but after that it now reverts to filename
completion similarly to when there was no specialized completion
support. getting this to work involved an unfortunate amount of bash
scripting even when making the bash completion function pretty general
...