capture_feats: working on using/debugging feature<->image space funcs.
also: add 'type' field to conv_op in conv support code and use it to
(hopefully) properly emulate caffe's input->output sizing for pooling
layers in cases that aren't exact. still TODO is to: (1) emit some
diagnostic when the pooling layers use 'implicit padding' (2) deal with
the need to potentially clip the support region of pooling pixels that
are near the + edge and use 'implicit' padding.