display_pil: hacky semi-broken state for computing recall from image-independant region boxes.
notes: intended to use boxes output when dump_rps=1 for
conv_pyra. pretty memory hungry. includes hack to clear some data
per-class during scoring, which conflicts with storing the matching
data for display. this is active when do_score=2 since we know we
won't try to do display. but, beware if you set do_score=1 with a
large # of images * classes * rp_boxes -- the memory usage can get to
> 32GB pretty fast. not that that's too useful of a mode to run in
anyway ...