rtc: working toward smem blocking.
for now, we're just conceptually reshaping blocks into (approx-square,
approx size 256) 2D tiling of threads rather than 1D. in parituclar,
this requires adjusting the # of threads per block to ensure there is
a full rectangle of tiles in each block -- for now we allow the # of
threads per block to go somewhat above 256 as needed. also, the change
can require increasing the number of blocks, since we may waste more
threads on the bottom border, since we can now waste whole tiles
instead of just partial tiles.. however, the kernel code isn't aware
of this restructing yet. since the # of blocks can only go up, and
that's the only thing the kernel might notice, the results should not
change (and they don't).

mwm