rtc: use NESI for nvrtc via has_conv_fwd_t base mode
nvrtc is the only submode of has conv_fwd_t currently, but we get
param support this way, and we maintain the seperation of the nvrtc
code from conv_pipe/run_cnet_t. obviously pulls in the complexity of
NESI, but otherwise code complexity impact is minimal, and argubally
for boda this is the standard/approved approach. of course it
highlights some missing boda/NESI help features for polymorphic
sub-fields ... err i think that doesn't work anyway.