caffeif: to improve robustness of timing caffe forward, do a (dummy) net->ForwardPrefilled() in caffe_create_net()
presumably this forced various lazy setup to occur, such as device
memory allocation, kernel setup, and who-knows what-all else. far from
foolproof, but it seems to make the first forward call take the same
time as later calls in both cudnn and non-cudnn modes. prior to this,
the first run would be much slower -- 2x in some cases, but could be
more or less; presumably it was a net-constant overhead.