nvrtc: remove restriction that out_chans must be evenly divisible by 8*N
where N was chan_tiles_per_block, ideally 16, and formed one side of
the block dims (so N*patch_tiles_per_block was ideally 256). also, try
to get threads_per_block as near as possible to the goal (usually
256), but don't go over. previously we'd get as near as possible but
ensure we were >= the value -- but this seems to hit some bad corner
cases for occupancy.