rtc: finished second pass at rework of s1conv. seems correct.
note that after the first rework pass, the results were identically
slightly wrong to before finishing the rework. the bias loading loop
was missing, and adding that fixed the results. it's possible that was
the only thing incorrect in the prior version. however, as
maybe-expected, the rework seems to speed-up and simpify things a bit
in terms of the input loading. overall, even with the tiny block
sizes, s1conv seems to have similar speed to the original conv
now. before the rework it was ~1.5X slower. still TODO is to actually
use the new compute structure for unrolling and saving redundant
loads.