rtc: multi-line support in s1conv. allows larger block sizes and/or smaller lines.
for existing trial cases, not much difference. but, we bump the block
sizes up from 32 to a maybe-more-sane 128 without any *loss* of perf
anyway. for the conv2-like-layers, we're probably above 50% of max
flops now. also, we're back to a state where it's unclear how to
further improve these cases and/or how they are currently limited.