rtc: avoid reading unused-but-OOB image data by guarding in loads on img_ix
note that s1conv() chooses to read from offset 0 of the 'in' pointer
rather than add conditionals to the actual loading code. regular
conv() will skip the load and use 0's for the values as if they were
padding. in both cases, the actual values should be unobservable, as
they only should effect outputs that are also out-of-bounds and thus
not written.