rtc: before trying to optimize k1conv in loading, add missing bounds check on in_chan
note that this is only needed when in_chan_tile (default 8) doesn't
evenly divide the # of chans, which is perhaps uncommon. if we try a
bit harder, we could only emit the check when needed.  it would also
be possible and reasonable to use padding to avoid such a
conditional. but for now it's one of several issue with the input
loading, so we defer any decisions/optimizations for now