rtc: broken checkpoint for k1conv() x,y -> pels conversion. looks like max 1 image/block is no good.
so, we probably need to allow >1 image block. but, it's still unclear
if partial images are sensible, or if it makes much
difference. presumably it would only matter when 1 image is near the
size of 1 block. if they are much larger then we split them into many
blocks, and only waste ~.5 blocks/image. if they are much smaller than
a block, we waste only ~.5 images worth.