rtc: first attempt at non-trival usage of in_xpose'ing for k1conv
by (generalized-)transposing the input (and adding padding), we can
simplify the bounds checking and k1conv and/or maybe improve the
memory access pattern when reading the input. note that, in exchange,
the output bounds checking is now somewhat more difficult -- it is
similar to the non-special-case conv where any two output pixels, even
within a thread, can span an abritrary dim in the output (or fall of
the end). however, on the whole, this change seems to allow us
(finallly) get some gains for all the k1conv cases, including when we
seem to need multiple images per block.