rtc: k1conv: choose tix_pels_tile_sz to be as-large-as-possible instead of always 8
in particular, cases where the tix_out_chan_tile_size is clamped to <
16 may now have a larger tix_pels_tile_sz. it's not clear if this is
good, bad, or unimportant. it seems like the current code does okay
over a pretty broad range of tile sizes and threads-per-block, which
is nice. but a max of ~128 TPB still seems best by a significant
margin.