rtc: add dummy stores to k1conv for xpose experiments.
note that writing out all the outputs in a trivial 'nice' (good access
pattern) way does pretty efficient, so that's nice. however, ideally
we'd like to output it the form we need it in next. in particular, the
next TODO is to try to see if we can do something good in the case
where the next conv is the same as the current one (i.e. as is the
case in between the size-1-kernel layers in NiN).