rtc: k1conv: add enable_write_xposed option.
when enabled, doesn't crash, but doesn't do anything useful yet
either. in this mode k1conv will attempt to produce output in the
correct layout to be directly used as input to another k1conv
layer. also, the explicit xpose prior to the next k1conv will be
ommited. this all seems to work, but the actual xposed-writing code
doesn't yet exist.