rtc: k1conv: add another (questionably useful) special case for filter load indexing
while working on the xposed-output-writing, i noticed some
sub-optimial ptx being generated for a particular case, which turned
out to be the case where (bix_out_chan_blk_sz==1). we can simplify an
indexing expressing in that case, allowing threadIdx.x to be pulled
into an filts_off_adj. as in similar prior cases, this seems to
convince/allow the compiler to generate better indexing (reg+const),
although it's a bit unclear why this is the case.