rtc: for 'fun', reorder the filts_xp dims, and see what needs to change.
some of the changes seem like the existing code was making bad
assumptions (in the code templates); the other changes seem mostly due
to the duplication of the nda_exprs inserts for filts_xp_ix in each of
the conv codegens. the template fixs seem good, but for the codegen
side we need to factor out the filter dims somehow. this is probably
related to the idea of storing the dims/exprs inside the node_info_t.