rtc: k1conv: first pass at xposed-writing
seems to be correct, and seems to avoid the need for any per-write
conditionals. note that it's possible garbage now propogates to places
that the old xpose would have set to zero, and that those values may
actually get used.  now, while the filts xpose doesn't 0-pad
currently, we zero-init all buffers. so, while it seems perhaps a bit
shady, we should be okay since at least the padded filters will be
zero and thus stamp out the garbage.

as expected, it's not faster than the out writing
method, but it does avoid the need for a subsequent
transpose. however, it's worth noting that it seems like it's actually
a bit *slower* than the original writing, and thus even with the xpose
removed the overall time may be higher. but, that's okay, since the
goal is to experiment with a better writing method, and we want to
start by trying to see if we can write things in the desired format
directly. failing that we can consider writing in other format (such
as the original format) and using an xpose. but we're hoping it
doesn't come to that!