rtc: k1conv: fix another trival sz<->dim typo in xposed-write. seems correct now.
note that this typo wasn't too easy to find, and required first
checking that all output slots were written correctly, and then
directly inspecting the outputs. it then seems that the issue was in
the out_smem reading, and upon re-working the indexing (and getting an
equivalent expression), the typo was noticed. i'm not a big fan of
these sz<->dim typos ... hmm. but not too bad overall, and would be
faster next time with the new cup_dump option adding during this
debugging phase.