rtc: tconv: multiple images per block (vertical stacking only)
this (in thoery, and maybe-hopefully in practice/measurment) removes 1
dimension of the overhead associated with small output images
(i.e. similar to the blocking size of ~8). for example, an output wih
27 lines was being rounded to 32 lines, which is about ~15% wasted
work. now, as long as there are a reasonable # of images in the batch,
this overhead is reduced: e.g. with 10 images, there are 270 total
lines, which is now >> than a blocking size of ~8.

the changes are not as simple as i'd like, but what did i expect?