rtc: refactoring events/profiling
working toward making profiling work in OpenCL. in this change, we
introduce a new (but unused) var-name-based profiling interface for
rtc, and partially impliment it (only for nvrtc). also, we split the
rfc function arguments into in/inout/out sets at the interface level,
although currently this split is unused in the backends.