prep work for backwards: always read train_val protobuf
for consistency, we're now trying to always read the train_val version
of the net on the thoery that it has a superset of the information in
the deploy net. but, we (still) convert it to a mostly-deploy format
(deplot, but also dropping any previously-read-but-mostly-ignored
Softmax layer (that we no longer expect to see in train_val style nets
anyway)). note that we were already dropping SoftmaxWithLoss, Data,
and Accuracy layers from train_val style nets. so this is more
consistent overall. reading deploy-style nets should still work but is
now untested.