more prep work for backwards: remedial-fwd: keep Softmax/SoftmaxWithLoss ...
.. but convert SoftmaxWithLoss->Softmax in the spirit of storing the
deploy-style net. Again, this is now consistent with the original
handling of deploy style nets. also, for now, stub out Softmax in
rtc_fwd to avoid failures when the Softmax layer is kept (the output
will be all zeros due to our zero-init of all vars), but unused. in
particular, in GoogLeNet, the Softmax's of the non-final classifiers
are kept/computed, but not used in any comparisons in testsing
(currently), since only a (single) output layer is compared. so we can
merrily compute it as zero without test failures.