change out_layer_name -> out_node_name
between layers-with-multiple-outputs and in-place-layers, we seems to
be in a pick-your-poison scenario. we can either be able to choose any
node, but *not* select a subset of in-place-ops that act on that node
OR we can choose any layer, but then can't pick an output. combined
with the new nodes that only exist after add_back_ops, and aren't in
the original prototxt at all, we have some issues. but, for now, we:
1) switch to using node names instead of layer names to name
outputs. this is hopefully okay, mostly compatible, and maybe we can
tweak the syntax to allow choosing subset of in-place-ops if need be
2) disable the error checking that the node must exist in the input
prototxt, so we can (try) to use nodes that only exist after
add_back_ops(). of course there's going to be some
caffe-compatibility-string-hackery at some point to make that work
overall, but here's hoping.

and, as usual, we consider our options for removing/flattening
in-place-nodes. we're probably not there yet, as we're still lacking
things like var lifetime analysis and whatnot that would seem to be
needed to make that go smoothly inside rtc_fwd.