add (optional) dim name to boda dims_t
note that we're also ditching the
single-shared-pointer-to-a-block-of-uint32_t based dims_t. it was
okay, but didn't really seem better than a simpler vector<> based
approach. now that we've got a string per-dim (even if it's usually
empty) i guess maybe we're not too worried about the size/perf of the
dims_t. hmm. anyway, for now, it seems like a unified dims_t is going
to be expedient, at least for experiments. we'll see how it goes.