rtc: semi-sideways process on rtc event handling. events+ipc now works.
replaced the pair of p_void's inside rtc_func_call_t with a single
(ipc-friendly) uint32_t. this is a bit cleaner iface wise and usage
wise as well -- the various backends create internal vectors to store
per-call data, and they can put whatever types they want in
it. however, in many ways, this isn't really too different from the
original p_void; backends could have used it to store a void * to
whatever they wanted already (although having two fields was
unneeded). and, now we don't get automatic freeing of the events. but,
the main reason to make this change now/finally is that it's easier to
deal with serializing a uint32_t as opposed to a p_void -- especially
if we were going to try to do some kind of shared-pointer-over-ipc
that actually tracked reference counts across the serialization
boundary (which we're not!).