rtc: remove ninfo, add oi->no_dims to replace it
note that oi->no_dims is used only in the is_conv case, and we're not
doing much different from before really. but hopefully this is moving
towards more ability to do sz->dims refactorings/changes (ideally
incrementally). as a side note, i think the existing ninfo handling
was incorrect, as ninfo was only set when an op was generated, not on
every call. perhaps in all tested cases, there was only one call of
each unique op with write_xformed=1?