rtc: redirect stats/var dumps to string (not stdout); add test for enable_stats=1
note that, currently, only test_compute will print the string returned
by the new get_info_log(), so ... that's unfortunate. just try to
remember if you enable stats/dump and nothing prints out, it's maybe
because you're running in a mode that doesn't print the info log (but
maybe it should).