rtc: working (backwards-ish?) toward single IO format for all convs
arguably, this is a feature regression, but for now it will make it
easier to port k1conv to the gen_call() flow. as a bonus, while it
does require implementing an 'untransform' kernel, the final result
shouldn't be any slow (maybe a bit faster). and, arguably, it's a bit
more sane in terms of input format handling. we should be able to
re-introduce multiple output (and/or input) formats later, but it's
worth noting it's not exactly clear how to extend the CUCL arg decl
syntax for that yet. maybe we want multiple sub-funcitons or the like
for that; then there'd be one sub-func for each input or output
format, and the right one would be called (probably with a static
conditional). but we need to make sure such sub-functions can somehow
work first ...