rtc: first pass at simple version of BckConv_in_grad (similar impl to spreading)
for one NiN case, seems to work up to pool3_grad_loss, which is as far
as we can go without adding the max case to spreading. we're also
still missing dropout, but maybe we can ignore that for a bit
longer. also, it's not clear how we're going to test the
{filts,biases}_grad_loss vars yet -- maybe we need to solve the
explicit filts/biases (FIXME_EFB) issue first. hmm.