conv_pyra: address an old FIXME by creating a temp net
this is to avoid needing to reset/alter in_num_imgs in the middle of
net setup, which in turn should allow us to simplify setup and help
with further sz->dims refactoring of conv_util/caffeif/caffepb.  note
that this is not well tested, and maybe not auto-tested at all, but
manually running the conv_pyra demo command seems to still work.