broken intermediate commit experimenting with refactoring massage_net_params
... by pushing it's functionality down into pipe_from_param and
caffe_fwd. this is moving moving toward helping fix the FIXME_EFB
issue as well as maybe making caffe_fwd 'cheat' less in terms of using
the input param directly as opposed to using a new param derived from
the pipe (which in turn is derived from the input param).