updates for 2015.11 version of caffe (7e40583db034).
the main changes are to (semi-)support arbirtrary-spatial-dim
convolutions. that is, we support the caffe.proto interface changes,
but only (as before) for the 2D case.

note that this update also means that caffe is using cudnnv3 by
default. a few tests have slightly differing results using cudnnv3
with the *old* caffe, and a few more differ when additionally using
the new caffe. all such tests are updated here. such changes are
expected, and the total # of differences and max_abs_diff is similar
in all cases, so there doesn't seem to be any concern.

note that to supress warning in the caffe headers and enable runtime
disabling of cudnn, it is reccomened to use the 'boda patches' version
of caffe, although this should not be strictly neccessary. but bear in
mind that is the normal development setup. at the time of this commit,
commit 52ce7186621cec6 on branch fix-sign-comp-in-hpps was used. this
branch should be pushed (and kept up to date) in the
git@github.com:moskewcz/caffe.git repo.