replace run_cnet_t in_sz,in_num_chan,in_num_imgs with in_dims.
yes, despite the # of files touched and the size of the diff, that's
pretty much all that was changed. various default shuffling and code
motion seemed to be needed/a-good-idea to make this work. as of this
commit, the tests pass. however, note that (as expected) some tests
were using improper (probably?) crop sizes of 227 when the net
.prototxt files specified 224 (which is probably the correct size to
use in those cases). in this commit, the tests affected continue to
override the crop size. in a following commit, we'll remove the
crop-size override, and update the values. test_dense_2 is tricky
because the behavior seems to change in that case so we'll need to
investigate that. also, it seems that maybe using a dims_t for in_dims
wasn't quite the right idea, and maybe we should switch it to a
map_str_uint32_t.