change type of newly-added dims_t in_dims to map_str_uint32_t
as mention before, this seems more sensible, since this field isn't
really a dims_t, but just a set of un-ordered overrides used to modify
the creation of a dims_t (by adjusting some of it's dims' sizes)
later.

for now, we leave the NESI support for dims_t present-but-unused.