remove now-unused-and-confusing num_imgs from flops.py
also, don't try to set it in rtc_fwd.cc's net output.  in flops.py, as
a consolation prize, print out the dims of an NDA named "data" if it
exists. it might be better to know what all the data-layer and/or
image source NDAs are (by annotation), and perhaps print the dims of
all those?