first pass at fixing FIXME_EFB's. the filts/biases are explicit now.
more work/cleanup may be needed, but it's not clear what exactly. this
change consisted of just visiting all the FIXME's, fixing them, and
then fixing all the other needed things until the tests passed. so,
there's probably various follow-on cleanups that will become clear
over time. the tests all pass, with the exception of test_stats_1,
which is updated here, since it now outputs stats for the conv1
filts/biases, and it's not clear that that's wrong/undesired. the
output ordering isn't too great, though, since alphabetical fails when
some node names are prefixes of others. hmm.