flops.py: code to (maybe?) better handle add_bck_ops=1 case
we now handle BckConv and move the parts of the old Convolution
handling that dealt with backwards in the BckConv section. however, we
choose to keep the statistics split up into backwards and forewards
categories as before, at least for compatibility and maybe because
that's more sensible. we now get at least basic perf. info for the
BckConv layers.