make mad_toler a param for test_compute; omit sds_str()/errs when no failure (mad or NaNs); enable grad tests
ommitting the sds_str() info and sample errors arguably makes the
tests have less coverage, but is more robust; small errors that are
always currently manually waived are now automatically
waived. howeve,r there is now no indication when there is a minor
varience in test results, which is unfortunate, because many changes
should not cause even a minor change in results, and such minro
changes might now be missed. however, for the grad tests in
particular, the caffe comparison results are non-deterministic in the
current setup, and thus *always* fail without this change. we might
perhaps want to re-enable the sds_str()/err prints for at least some
tests. hmm.