rtc: rather than refactor filts_smem loading in s1conv, remove it wholesale.
IIRC, s1conv is still better than tconv in some cases, but not by a
lot. for now, to simplify things, we'll just remove it. the main issue
is that since it's not tested, we'd really need to add more tests
before we'd have any confidence that any modifications to it were
correct. and, who knows if the current state is even correct.