lrn: add a mode to match the behavior of caffe LRN
see the notes in lrn.cucl; this allows exact agreement up to
norm2/pool2 in alexnet_ng_conv with cuDNN disabled. however, at conv3
there are still significant variances (~1e-4 max abs diff) that should
be investigated.

this change was part of an attempt to get better agreement on conv5
before debugging ipconv. however, it seems that due to other issues
(i.e. conv3), this doesn't end up helping us. so for now, it's
disabled.

as usual, this brings up the issue that we probably should have more
unit-y tests for invididual layers to help with cases like this. but
for now, the results are hopefully okay enough to proceed. we can
always debug detailed precision errors later, right? ;)