rtc: more work on ipconv; has OOB reads, maybe other bugs, but otherwise works
the known OOB reads have FIXMEs, and there aren't any other known bugs
or OOB accesses, although it wouldn't be too surprising if there
were. we could fix the issues now, but the plan is to ponder the best
overall solution, bearing in mind the general issue of having a nice
shared data layout (to the degree possible) for all the pre-existing
conv variants, ipconv, maybe new conv variants, and especailly for the
BckConv problems sizes. the obvious candidation is chans-innermost,
with some fixed padding (i.e. enforce (chans%32)==0).