rtc: add incomplete bwai codegen for testing/tuning
some initial work toward having some form of basic sgemm and/or other
simple benchmarks for testing bandwidth/compute/AI. might be also
useful for intial demontrations/experiments with autotuning. however,
we really need some additional front-end work to be able to use this
or do tuning experiments. in particular, we perhaps need a more
general and/or non-cnn front end, and we maybe need to split the rtc
laeyr into a 'general' layer + a 'cnn-specific' layer. much of the
recent refactoring and cleanups should make this generally more
feasible. however, in some cases, some flows may be too specific to
cnn processing, and some simplifications / factorings might need to be
backed out.