rtc: stub out bconv (bck version of 'regular' conv).
on consideration, it's probably better to start with the bconv rather
than btconv. it seems a bit trivial to share the init() blocking setup
code between bck and fwd, so we're going to clone it and strip it down
for just bconv (hopefully factoring out the basic tiling calculation
part).