rtc: bconv: rework bck_in_pad / size calculation; fix out->in calc for >1 kernel size
believe it or not, this version of the bck size/padding calc makes
more sense to me. it's not clear the original was wrong. in this
version, we also calculate bck_pad_in_off, but that's just
(kern_sz-1)*stride, which we could have calculated before. anyway, for
now we're claiming the correct output->input space mapping is:
(nopad_out*stride) + bck_pad_in_off - in_pad + sub_kern_off. note that
we could fold the "- in_pad" into the bck_pad_in_off and maybe rename
it, but again (believe it or not) this way seems easier to
undestand. also not that, in general, the sub_kern_off / stride>1
case(s) are not yet tested and are probably wrong.