add enable_bconv=1 and alexnet_ng_conv_nd_nl gradient tests
note: we're still missing concat/split, LRN support, and dropout
support, which we need for testing of 'regular' alexnet_ng_conv and
googlenet_conv. the xxx_nd_nl variant of alexnet has no LRN or
dropout. we can make a similar variant of googlenet one we have just
split support, which should be enough to test that. but first we need
a test that actually uses stride>1 for backprop. and further we may
work on the filts/biases grad before we that.