change default models_dir to point to boda_dir/nets; add alt_models_dir (used only by run_cnet_t)
the general idea is to move toward a boda working copy being more self
contained. since we're putting the net defs in the repo now, it makes
sense to actually read them from the repo/working copy for the tests
and general usage.

note that one new net used by testing that wasn't in the repo is added
here; case in point for this change?

run_cnet_t will now look for the model paramters file in
alt_trained_fn if it is not found in trained_fn. alt_trained_fn
defaults to "%(alt_model_dir)/%(model_name)/best.caffemodel"

for now, other users of must_load_binary_proto() don't support the
alternate models directory/filename. they probably should, but they're
not tested and it's not 100% clear what we really want anyway.

also, add expanded filename to some error printout, since it's nuts
for it not to be there. it probably need to go in more places, but
there are the orignial testing-related issue why we didn't do that to
consider.