experimenting with uninterpreted/unchecked parameters for conv_op_t
starting with just the lrn parameters. as it stands, while the
defaults and list of param names is stored in the coi, it is
ununsed. so, all needed per-type parameters must be set, and the only
checking that they are set is if/when they are used in the compute
backends. the tests pass, so i guess nobody ever uses the boda-level
defaults for LRN layers ...