conv_util: make in_pad / eff_tot_pad be u32_pt_t instead of u32_box_t
or, drop support for different padding on +- edges of any particular
dim. we never used that support, and now it seems like it's a bit in
the way. now we can treat kern_sz, stride, and in_pad all as
per-spatial-dim, which means they are currently u32_pt_t, but could
now also be dims_t.