rtc: working toward removing cop from op_info_t
this creates a clean seperation between the conv_util and rtc_fwd
layers. however, note that there's a fair amount of duplication
between the two levels, so it's not totally clear this is the greatest
idea. however, it seems like a reasonable step toward the further
separation between the post-init operation and the
individual-functions/variant level(s?).