conv_util/flops.py: cleanups/refactorings of dump_ops
working toward a bit more sanity/completeness in the boda pyIR
format. also, working toward the removal of the global 'net' in
flops.py inside the ops. it should be removable now, but it's still
there for now. note that the plan currently remains for the pyIR to
reference 'net' directly for now; we just specifically trying to
remove the usage of 'global net' from inside the ops.