nets: add modified version of firenet8 with conv_final padding removed
this 'bad'/weird/unneeded padding causes two issues: (1) can't use
k1conv with padding -- this is nonfatal but irritating and makes
profiling arguable less meaningfull, since conv_final will be slower
than 'it should be' (2) we end up with negative padding in BckConv for
conv_final. this is currently fatal/unsupported, and maybe we do want
to handle it, but for now it seems better to avoid it.