rtc: remove positional arg handling
had to change the various remaining pos arg usages in rtc_fwd to use
by-name args. in particular, note that the various in/filts
transformation function had thier arg names changes to match the names
in oi->ref_dims. it's not super clear if this is best, and the
gen_call() parts for those funcs got a bit uglier, but it seems okay
for now.