port test_dense and test_upsamp to use MAD_FAILS/ALL_IS_WELL system
instead of always printing differences. these results aren't currently
non-deterministic, but we're experimenting with changes/enhancements
to the string from of ssds_diff_t, so it's unconvenient to have them
always printing bunches of ssds strings out. test_dense was all 0
errors, so we set mad_toler == 1e-99 and say that's basically
equivalent and cleaner. test_upsamp has larger absolute differences,
from around 1e-5 to 9e-5, so we set the toler == 1e-4, and hope we're
not losing too much of interest by dropping the check that the results
be exactly equal. the results should be more robust now, which is a
nice consolation.

there's also some preliminary/experimental changes to ssds_diff_t in
this commit, but they are unused and probably mostly wrong/useless.