testing: make test_compute test all nodes, not just outputs
this is imperfect, since not all conv_pipe nodes need exist in the
caffe net, and it involves copying out various 'const' nodes such as
the filters and biases ... which might be especially sketchy if they
are reused. however, it seems currently like this is not the case so
all should be well. hmm.

also, testing all nodes required quite a bit of per-node-tolerance
specifications, especially for alexnet and googlenet, probably due to
LRN issues. see comments in test_cmds.xml and lrn.cucl