rtc: first pass at full (N-input) reduce support
it seems a bit ugly/experimental, but it should be correct/okay. may
need/want to revisit. also, note that the basic googlenet tests still
seems to have 'too much' error, so they are still disabled pending a
bit more investigation. maybe there's some issue with multiple
classifiers? anyway, we'll make some more attempts to fix it, and then
work on dropout. after that the plan is go back to per-op
profiling/testing, which hopefully will highlight/shake-out various
remaining issues with accuracy (although it's worth noting that many
error are more related to the combination of operations than to the
individual operations).