nets: add no-LRN version of googlenet for debugging
not that LRN is neccessarily wrong, but the numerical differences
between implimentations (wrong or not) make debugging harded. with
this version, it seems icp8_out_grad_loss is the first layer (in
compute order) with 'bad' errors. note that cls1_pool and cls2_pool
have pretty high absolute errors (even just in fwd) of around ~1e-3,
but they also have large values (~1e3), so that's probably fine. one
of the rare cases where the max-relative-diff seems to work and give a
valid result (~1e-7) ... sigh. anyway, icp8_out_grad_loss being bad
points to the bck versions of reduce and concat, as expected. however,
to process, it seems best to try to actually compare the reduce inputs
for icp8_out_grad_loss, which means figuring out how to get them out
of caffe ...