conv_util: questionable hack to make var/blob names agree between caffe and boda
also, enable testing of said variables by removing the
hack/conditional that was previously disabling testing for them. seems
to work okay for googlenet, but seem brittle/incomplete. in short, it
assumes all such nodes are the output of Concat layers in caffe, where
the layer name and output blob name are equal, which is clearly not
general. we may disable, remove, or make-optional this code when we're
done debugging googlenet. on a related note, it looks we'll need to
emulate the 'pick first' behavior of caffe/cuDNN's
backwards-max-pooling.