conv_util: dropout support, fwd and bck. finally!
testing requires caffe with support for hack to allow deterministic
dropout. my github fork on the usual fix-sign-comp-in-hpps should
generally be the right version to use and this case is no exception.

note that there's still a few issues to track down. in particular,
there are failures (too many zeros after bck pooling) when using
cuDNN, but it's unclear who's right/wrong yet. it should be noted that
the gradients are pretty sparse already in the test-case in question,
and that it uses a non-zero pooling stride which makes things more
interesting.