rtc: oops, dropout should refer to inout in it's IX decl, not out
note that this wasn't a correctness issue, but turned up as an issue
as part of the 'reduced' function signature refactoring. IX decls
should only refer to actual args, not random unused stuff in the
'full' ref_dims, so this was always illegal, but of course correct
since inout has the same dims and in and out. and, again, stupid
in-place operations ... sigh.