rtc: various refactoring to make rtc_func_sig_t a NESI class
hopefully this is a good idea for reading/writing set of these. we
needed to add support for map_str_dims_t to NESI. also, for this to
work, it looks like we need to change the default dims_t str()
behavior to be the param format, which isn't done in this commit yet.