rtc_prof: add alloc + running calls. working for tiny example.
on large example, next sticking point is functions with u32 args, in
particular those with var-sized inputs, which show up as zero size in
the signature, and hence are hard to allocate. it's unclear exactly
how to deal with them. of course, they maybe aren't particularly
important for profiling, but we need to do somthing with them ... hmm.