updates for 2016.02.27 caffe: update caffe.proto, that's it
needed to slightly increase tolerance on nin tests, probably due to
the new caffe using cuDNN for avg pooling. prior to this, we had
patched caffe to disable cuDNN pooling altogether (due to incorrect
behavior of cuDNN pooling when preceded by in-place layers like
dropout, see prior comments), but in current caffe there is a more
limited workaround that only disables cuDNN pooling for the
max-pooling case.

note that the updated caffe.proto just adds new features that we
ignore/don't support, so no code changes were needed. in particular,
it adds a 'dialation' parameter to conv layers, which we ignore, but
should probably at least error-check is not set. it's not clear if
caffe allows stride *and* dialation at the same time, but i'd guess in
general that would be weird. anyway, dialation is similar / the same
as the way we handle stride for backwards convolution, so maybe we
could support it at some point.