replace usage of conv_op stride with dims_vals["stride"]
still TODO is similar treatment of kern_sz and in_pad. while this
change makes it possbile for stride to be ND (instead of just 2D)
most/many users still only support the 2D case. in general, this is
indicated by them calling the stride() function, which checks that the
stride is 2D and returns a u32_pt_t (the old type of stride). this
seems fine/good transitionaly at least. our main current goal with
this refactoring is to harmonize dims_vals/str_vals across usages, not
to enable ND support for stride/kern_sz/in_pad.